+++ src/main/java/net/djahmo/bakeddelight/custom/block/crop/ZucchiniCropBlock.java
@@ -1,0 +1,43 @@
+package net.djahmo.bakeddelight.custom.block.crop;
+
+import net.djahmo.bakeddelight.registry.ModItems;
+import net.minecraft.world.level.ItemLike;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.CropBlock;
+import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.StateDefinition;
+import net.minecraft.world.level.block.state.properties.BlockStateProperties;
+import net.minecraft.world.level.block.state.properties.IntegerProperty;
+
+public class ZucchiniCropBlock extends CropBlock {
+
+    public static final int MAX_AGE = 5;
+    public static final IntegerProperty AGE = BlockStateProperties.AGE_5;
+
+    public ZucchiniCropBlock(Properties pProperties) {
+        super(pProperties);
+    }
+
+    @Override
+    protected ItemLike getBaseSeedId() {
+        return ModItems.ZUCCHINI_SEED.get();
+    }
+
+    @Override
+    protected void createBlockStateDefinition(StateDefinition.Builder<Block, BlockState> pBuilder) {
+        pBuilder.add(AGE);
+    }
+
+    @Override
+    public IntegerProperty getAgeProperty() {
+        return AGE;
+    }
+
+    @Override
+    public int getMaxAge() {
+        return MAX_AGE;
+    }
+
+
+}
+