+++ src/main/java/net/djahmo/bakeddelight/custom/datagen/loot/ModBlockLootTables.java
@@ -1,6 +1,7 @@
 package net.djahmo.bakeddelight.custom.datagen.loot;
 
 import net.djahmo.bakeddelight.custom.block.crop.EggplantCropBlock;
+import net.djahmo.bakeddelight.custom.block.crop.ZucchiniCropBlock;
 import net.djahmo.bakeddelight.registry.ModBlocks;
 import net.djahmo.bakeddelight.registry.ModItems;
 import net.minecraft.advancements.critereon.StatePropertiesPredicate;
@@ -29,16 +30,23 @@ public class ModBlockLootTables extends BlockLootSubProvider {
         this.dropSelf(ModBlocks.UNCOOKED_GRATIN_DISH.get());
         this.dropOther(ModBlocks.COOKED_GRATIN_DISH.get(), ModBlocks.BAKING_DISH.get());
 
-
         this.dropSelf(ModBlocks.UNCOOKED_MOUSSAKA_DISH.get());
         this.dropOther(ModBlocks.COOKED_MOUSSAKA_DISH.get(), ModBlocks.BAKING_DISH.get());
 
+        this.dropSelf(ModBlocks.UNCOOKED_TIAN_DISH.get());
+        this.dropOther(ModBlocks.COOKED_TIAN_DISH.get(), ModBlocks.BAKING_DISH.get());
 
-        LootItemCondition.Builder dropCondition = LootItemBlockStatePropertyCondition
+
+        LootItemCondition.Builder dropConditionEggplant = LootItemBlockStatePropertyCondition
                 .hasBlockStateProperties(ModBlocks.EGGPLANT_CROP.get())
                 .setProperties(StatePropertiesPredicate.Builder.properties().hasProperty(EggplantCropBlock.AGE, 5));
 
-        this.add(ModBlocks.EGGPLANT_CROP.get(), createCropDrops(ModBlocks.EGGPLANT_CROP.get(), ModItems.EGGPLANT.get(), ModItems.EGGPLANT_SEED.get(), dropCondition));
+        this.add(ModBlocks.EGGPLANT_CROP.get(), createCropDrops(ModBlocks.EGGPLANT_CROP.get(), ModItems.EGGPLANT.get(), ModItems.EGGPLANT_SEED.get(), dropConditionEggplant));
+
+        LootItemCondition.Builder dropConditionZucchini = LootItemBlockStatePropertyCondition
+                .hasBlockStateProperties(ModBlocks.ZUCCHINI_CROP.get())
+                .setProperties(StatePropertiesPredicate.Builder.properties().hasProperty(ZucchiniCropBlock.AGE, 5));
+        this.add(ModBlocks.ZUCCHINI_CROP.get(), createCropDrops(ModBlocks.ZUCCHINI_CROP.get(), ModItems.ZUCCHINI.get(), ModItems.ZUCCHINI_SEED.get(), dropConditionZucchini));
     }
 
     @Override