+++ src/main/java/net/djahmo/bakeddelight/registry/ModBlocks.java
@@ -6,6 +6,7 @@ import net.djahmo.bakeddelight.custom.block.BakingCookedDishBlock;
 import net.djahmo.bakeddelight.custom.block.BakingDishBlock;
 import net.djahmo.bakeddelight.custom.block.BakingUncookedDishBlock;
 import net.djahmo.bakeddelight.custom.block.crop.EggplantCropBlock;
+import net.djahmo.bakeddelight.custom.block.crop.ZucchiniCropBlock;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.level.block.Block;
@@ -21,19 +22,17 @@ import java.util.function.Supplier;
 
 public class ModBlocks {
     public static final DeferredRegister<Block> BLOCKS;
-    public static final RegistryObject<Block> BAKING_DISH, UNCOOKED_LASAGNA_DISH, COOKED_LASAGNA_DISH, UNCOOKED_GRATIN_DISH, COOKED_GRATIN_DISH, UNCOOKED_MOUSSAKA_DISH, COOKED_MOUSSAKA_DISH;
-    public static final RegistryObject<Block> EGGPLANT_CROP;
-
-    private static int nextDishId = 0;
+    public static final RegistryObject<Block> BAKING_DISH, UNCOOKED_LASAGNA_DISH, COOKED_LASAGNA_DISH, UNCOOKED_GRATIN_DISH, COOKED_GRATIN_DISH, UNCOOKED_MOUSSAKA_DISH, COOKED_MOUSSAKA_DISH, UNCOOKED_TIAN_DISH, COOKED_TIAN_DISH;
+    public static final RegistryObject<Block> EGGPLANT_CROP, ZUCCHINI_CROP;
 
     public static void register(IEventBus eventBus) {
         BLOCKS.register(eventBus);
     }
 
-    public static RegistryObject<Block> setBlock(String name, Supplier<Block> block, boolean cookedRecipe, Integer maxStackSize) {
+    public static RegistryObject<Block> setBlock(String name, Supplier<Block> block, boolean cookedRecipe, Integer maxStackSize, Integer id) {
         RegistryObject<Block> temp = BLOCKS.register(name, block);
         setItemBlock(name, temp, maxStackSize);
-        if(cookedRecipe)  BackingDishTypeCollection.registerDishType(name, nextDishId++);
+        if(cookedRecipe)  BackingDishTypeCollection.registerDishType(name, id);
         return temp;
     }
 
@@ -44,17 +43,21 @@ public class ModBlocks {
     static {
         BLOCKS = DeferredRegister.create(ForgeRegistries.BLOCKS, BakedDelight.MODID);
 
-        BAKING_DISH = setBlock("baking_dish", () -> new BakingDishBlock(BlockBehaviour.Properties.copy(Blocks.STONE).sound(SoundType.STONE).strength(0.4F)), false, 16);
+        BAKING_DISH = setBlock("baking_dish", () -> new BakingDishBlock(BlockBehaviour.Properties.copy(Blocks.DIRT).sound(SoundType.STONE).strength(0.4F)), false, 16, 0);
 
-        UNCOOKED_LASAGNA_DISH = setBlock("uncooked_lasagna_dish", () -> new BakingUncookedDishBlock(BlockBehaviour.Properties.copy(Blocks.STONE).sound(SoundType.STONE).strength(0.4F)), false, 1);
-        COOKED_LASAGNA_DISH = setBlock("cooked_lasagna_dish", () -> new BakingCookedDishBlock(BlockBehaviour.Properties.copy(Blocks.STONE).sound(SoundType.STONE).strength(0.4F), ModItems.LASAGNA_SLICE), true, 1);
+        UNCOOKED_LASAGNA_DISH = setBlock("uncooked_lasagna_dish", () -> new BakingUncookedDishBlock(BlockBehaviour.Properties.copy(Blocks.DIRT).sound(SoundType.STONE).strength(0.4F)), false, 1, 1);
+        COOKED_LASAGNA_DISH = setBlock("cooked_lasagna_dish", () -> new BakingCookedDishBlock(BlockBehaviour.Properties.copy(Blocks.DIRT).sound(SoundType.STONE).strength(0.4F), ModItems.LASAGNA_SLICE), true, 1, 1);
 
-        UNCOOKED_GRATIN_DISH = setBlock("uncooked_gratin_dish", () -> new BakingUncookedDishBlock(BlockBehaviour.Properties.copy(Blocks.STONE).sound(SoundType.STONE).strength(0.4F)), false, 1);
-        COOKED_GRATIN_DISH = setBlock("cooked_gratin_dish", () -> new BakingCookedDishBlock(BlockBehaviour.Properties.copy(Blocks.STONE).sound(SoundType.STONE).strength(0.4F), ModItems.GRATIN_SLICE), true, 1);
+        UNCOOKED_GRATIN_DISH = setBlock("uncooked_gratin_dish", () -> new BakingUncookedDishBlock(BlockBehaviour.Properties.copy(Blocks.DIRT).sound(SoundType.STONE).strength(0.4F)), false, 1, 2);
+        COOKED_GRATIN_DISH = setBlock("cooked_gratin_dish", () -> new BakingCookedDishBlock(BlockBehaviour.Properties.copy(Blocks.DIRT).sound(SoundType.STONE).strength(0.4F), ModItems.GRATIN_SLICE), true, 1, 2);
 
-        UNCOOKED_MOUSSAKA_DISH = setBlock("uncooked_moussaka_dish", () -> new BakingUncookedDishBlock(BlockBehaviour.Properties.copy(Blocks.STONE).sound(SoundType.STONE).strength(0.4F)), false, 1);
-        COOKED_MOUSSAKA_DISH = setBlock("cooked_moussaka_dish", () -> new BakingCookedDishBlock(BlockBehaviour.Properties.copy(Blocks.STONE).sound(SoundType.STONE).strength(0.4F), ModItems.MOUSSAKA_SLICE), true, 1);
+        UNCOOKED_MOUSSAKA_DISH = setBlock("uncooked_moussaka_dish", () -> new BakingUncookedDishBlock(BlockBehaviour.Properties.copy(Blocks.DIRT).sound(SoundType.STONE).strength(0.4F)), false, 1, 3);
+        COOKED_MOUSSAKA_DISH = setBlock("cooked_moussaka_dish", () -> new BakingCookedDishBlock(BlockBehaviour.Properties.copy(Blocks.DIRT).sound(SoundType.STONE).strength(0.4F), ModItems.MOUSSAKA_SLICE), true, 1, 3);
+
+        UNCOOKED_TIAN_DISH = setBlock("uncooked_tian_dish", () -> new BakingUncookedDishBlock(BlockBehaviour.Properties.copy(Blocks.DIRT).sound(SoundType.STONE).strength(0.4F)), false, 1, 4);
+        COOKED_TIAN_DISH = setBlock("cooked_tian_dish", () -> new BakingCookedDishBlock(BlockBehaviour.Properties.copy(Blocks.DIRT).sound(SoundType.STONE).strength(0.4F), ModItems.TIAN_SLICE), true, 1, 4);
 
         EGGPLANT_CROP = BLOCKS.register("eggplant_crop", () -> new EggplantCropBlock(BlockBehaviour.Properties.copy(Blocks.WHEAT).sound(SoundType.CROP).noCollission().noOcclusion()));
+        ZUCCHINI_CROP = BLOCKS.register("zucchini_crop", () -> new ZucchiniCropBlock(BlockBehaviour.Properties.copy(Blocks.WHEAT).sound(SoundType.CROP).noCollission().noOcclusion()));
     }
 }