+++ src/main/java/net/djahmo/bakeddelight/registry/ModCreativeTabs.java
@@ -33,6 +33,8 @@ public class ModCreativeTabs {
                             pOutput.accept(ModBlocks.COOKED_GRATIN_DISH.get());
                             pOutput.accept(ModBlocks.UNCOOKED_MOUSSAKA_DISH.get());
                             pOutput.accept(ModBlocks.COOKED_MOUSSAKA_DISH.get());
+                            pOutput.accept(ModBlocks.UNCOOKED_TIAN_DISH.get());
+                            pOutput.accept(ModBlocks.COOKED_TIAN_DISH.get());
 
                             //Items
                             pOutput.accept(ModItems.GROUND_MEAT.get());
@@ -40,6 +42,8 @@ public class ModCreativeTabs {
                             pOutput.accept(ModItems.LASAGNA_PASTA.get());
                             pOutput.accept(ModItems.SLICED_POTATO.get());
                             pOutput.accept(ModItems.SLICED_EGGPLANT.get());
+                            pOutput.accept(ModItems.SLICED_ZUCCHINI.get());
+                            pOutput.accept(ModItems.SLICED_TOMATO.get());
                             pOutput.accept(ModItems.CHEESE.get());
                             pOutput.accept(ModItems.BACON_BITS.get());
                             pOutput.accept(ModItems.COOKED_BACON_BITS.get());
@@ -47,10 +51,13 @@ public class ModCreativeTabs {
                             pOutput.accept(ModItems.BACON_CREAM.get());
                             pOutput.accept(ModItems.EGGPLANT_SEED.get());
                             pOutput.accept(ModItems.EGGPLANT.get());
+                            pOutput.accept(ModItems.ZUCCHINI_SEED.get());
+                            pOutput.accept(ModItems.ZUCCHINI.get());
 
                             pOutput.accept(ModItems.LASAGNA_SLICE.get());
                             pOutput.accept(ModItems.GRATIN_SLICE.get());
                             pOutput.accept(ModItems.MOUSSAKA_SLICE.get());
+                            pOutput.accept(ModItems.TIAN_SLICE.get());
                             //Tools
                             pOutput.accept(ModItems.ROLLING_PIN.get());