+++ src/main/java/net/djahmo/bakeddelight/registry/ModLootModifiers.java
@@ -1,0 +1,26 @@
+package net.djahmo.bakeddelight.registry;
+
+import com.mojang.serialization.Codec;
+import net.djahmo.bakeddelight.BakedDelight;
+import net.djahmo.bakeddelight.custom.loot.AddItemModifier;
+import net.minecraftforge.common.loot.IGlobalLootModifier;
+import net.minecraftforge.eventbus.api.IEventBus;
+import net.minecraftforge.registries.DeferredRegister;
+import net.minecraftforge.registries.ForgeRegistries;
+import net.minecraftforge.registries.RegistryObject;
+
+public class ModLootModifiers {
+    public static final DeferredRegister<Codec<? extends IGlobalLootModifier>> LOOT_MODIFIER_SERIALIZERS;
+
+    public static final RegistryObject<Codec<? extends IGlobalLootModifier>> ADD_ITEM;
+
+    public static void register(IEventBus eventBus) {
+        LOOT_MODIFIER_SERIALIZERS.register(eventBus);
+    }
+
+    static {
+        LOOT_MODIFIER_SERIALIZERS = DeferredRegister.create(ForgeRegistries.Keys.GLOBAL_LOOT_MODIFIER_SERIALIZERS, BakedDelight.MODID);
+
+        ADD_ITEM = LOOT_MODIFIER_SERIALIZERS.register("add_item", AddItemModifier.CODEC);
+    }
+}