+++ src/main/java/net/djahmo/bakeddelight/utils/Function.java
@@ -5,10 +5,16 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.ItemLike;
 
 public class Function {
+
     public static ResourceLocation getResourceLocation(ItemLike item) {
-        return new ResourceLocation(BakedDelight.MODID, item.asItem().getDescriptionId().split("\\.")[2]);
+        return new ResourceLocation(BakedDelight.MODID, getItemName(item));
     }
+
     public static ResourceLocation getResourceLocation(ItemLike item, String postfix) {
-        return new ResourceLocation(BakedDelight.MODID, item.asItem().getDescriptionId().split("\\.")[2] + postfix);
+        return new ResourceLocation(BakedDelight.MODID, getItemName(item) + postfix);
+    }
+
+    public static String getItemName(ItemLike item) {
+        return item.asItem().getDescriptionId().split("\\.")[2];
     }
 }