+++ changelog.md
@@ -1,3 +1,14 @@
+# Version 0.1.0
+
+- Add Tian recipe
+- Add Zucchini Crop
+- Fix bug on JEI integration
+- Block model are now generated
+- Blockstate are now generated
+- All recipe are now generated
+- Add seed recipe and drop from grass
+- Add craft for chef suit
+
 # Version 0.0.2
 
 - Add Moussaka recipe